Show current tab visible on tab bar even if there are a lot of open tabs
Previously, if you had a lot of tabs and/or tabs with long names,
the current tab might not show up in the tab bar (if its position
in the tab name string was more than $self->ncol chars from the
beginning).  Changes below ensure that the name of the current tab
is always visible.  Arguably it would be nice to have indications
at each end showing whether there are more tabs in that direction.

Also, when printing the tab bar, limit the length of the tab
names so that one tab name doesn't occupy the whole tab bar.

[mina86@mina86.com: tiny refactoring]