Add documentation
Convert the long unstructured comment at the beginning of the tabbedex
file into a POD documentation which can be converted into a man page or
an HTML document.  Create Makefile rules to do both.

Also move list of authors to separate AUTHORS file since there’s no
longer an abvious place to keep it in the source file itself.