Fix broken tab bar when reopen-on-close and last tab is closed
When last tab is closed with reopen-on-close option in effect, tab
bar becomes mangled showing two tabs with the second one as active.
(Minimising and then showing the window again solves the issue but,
it’s clearly not a desired behaviour).  Fix the bug by making sure
refreshing happens at correct time.

Reported-by: David Koňařík <dvd.kon@gmail.com>