Don’t call make_current at the end of move_tab
move_tab does not change the current tab therefore there is no need for
it to call make_current method.  The actual desired side effect of calling
that method is that it’ll unconditionally call $self->refresh.  Remove the
indirection and call $self->refresh directly in move_tab.

While dealing with that function, convert it to use early return so that
one level of indentation is removed and thus hopefully function made more
readable.