Rework tabbar rendering code
First of all, when not all tabs can fit in the tabbar, make sure that
the current one *and* the one following it are displayed.  This makes
it clear that there are still tabs on the right of the active one.

Second of all, utilise as much space as possible for displaying tab
names instead of creating the same width limit for all tabs (even
those whose names are nowhere near the limit).

For example, previously if there were two tabs each tab’s name would be
truncated to fill at most half the tab bar.  If one of the tabs had
one-letter name, the other’s name would be truncated at half the tabbar.
Currently, the first tab will be displayed in full and the other will be
limited to occupy the rest of the tab bar.

The result is that in more cases tab names will be displayed in full.