Fix set_tab_name OSC not working properly while renaming tab
If set_tab_name OSC is used in a non-active tab, let’s call it
$foo, while the active tab, let’s call it $cur, is being renamed,
the code will assume that $foo is being renamed.  This results in
the new name (coming from the OSC) to be stored in $foo->{old_name}
which is silently ignored.

Fix so that the code correctly checks whether the tab where OSC is
issued is being renamed or not.