Move tab renaming code to tab package
Isolate code which accesses and thus needs to understand tab’s old_name
property to a few clustered functions in the tab package.

Introduce a new is_being_renamed method so other parts of the code can
test for that without having to know the implementation of that check.

With that, observe that only current tab can be renamed thus
is_inputing_name property is not necessary as {cur}->is_being_renamed
can be checked innstead.