Use single array to store all rends used by the extension
Collapse rs_tabbar, rs_tab and rs_title into a single array.
rs_tab is already one and its first element is implicitly the same
as rs_tabbar pso it’s enough to just add rs_title at the end of it
and things will be trivial to migrate.

This simplifies things slightly because now rends can be easily
referenced by a single number.  It is also a small refactoring
which will make the next commit easier.

While doing that, use title’s rend to fill in any left-over space
in the tabbar if title was displayed.  This way the right padding’s
bacgkround will match title’s.