Get rid of update_autohide
First of all, get rid of the ‘$reconfigure’ argument to update_autohide
method.  With that, reconfiguratino will happen every time tabbar’s
visibility is toggled.  There was just one place in the code where this
argument was non-true (in tab_destroy method) so while we might be doing
more reconfigurations than previously, the adde cognitive load is not
worth it.

Second of all, observe that each call to update_autohide is followed by
call to refresh (even in tab_destroy where refresh is either called
directly or through make_current).  Based on that, simply inline the
former in the latter to simplify the code.