Rename $self to $root
Use a single consistant name for the main terminal window (which
displayed the tabbar and held the actual tabs) so that the same
code can be read easily without having to think of context where
it’s in.

In some places ‘parent’ was used as the name but that had a conflict
with parent method which returns terminal’s X window.  ‘root’ seems
more appropriate name.