Don’t upidate last_activity in current tab
last_activity property is ignored in current tab so there is no point
in updating it when active tab gets new content.  As a consequence,
it’s no longer necessary to check if tab is not the current one in
tab_activity_mark since last_activity being undefined covers tab
being the current one.