Introduce pass-through package acting as the plugin’s main entry point
Create a single entry point for master and slave tabs so that from urxvt’s
point of view the same plugin is loaded each time.  This will make its
name (e.g. ‘$self->{_name}’ property but also other references) behave the
same way regardless whether given terminal is root or not.

This is necessary to make actions work since in their case urxvt does the
string matching and previously it would treat tabs as if their extension
name was ‘tabbedex::tab’.  This could never match any action.

This is where ‘_on’ subroutine defined in the previous commit comes in
handy as without explicit enabling hook, in this new scheme, urxvt would
not enable any hooks since it would look for handles in the wrong package.

Since the tabs now use the same extension name as the root, black listing
of ‘tabbedex’ extension is no longer necessary.

While at it, use ‘tabbedx-root-tab’ property in the terminal to store
reference to the master tab and pull that reference to tab’s hash on
init.  Changed name makes it very unlikely that any clash will happen
and pulling the reference to tab’s object saves one dereference.