Add support for actions
Since urxvt 9.21+ user commands are deprecated in favour of actions.
The advantage of actions is that they are routed directly to the
extension they are directed at rather than being sent to all plugins.

Because tabbedex uses separate package for slave tabs handling actions
is not straightforward but with introduction of pass-through package
in commit [5225f01a35: ‘Introduce pass-through package acting as the
plugin’s main entry point’] comes in handry.

Compatiblity code makes tabbedex work in older urxvts as well though
that code has not been tested.