Fixes to release script
a) Fix argument parsing.  ‘for arg; do’ or ‘shift’; but not both.
Went with the latter.

b) Accept refspec of a new master and experimental.  Two positional
arguments are now supported.  The first specifeis commit to cut the
release from and the second refspec of experimental branch.

c) Push created tag as well as master and experimental branches to
the origin.  With default settings, pushing master and experimental
is redundant but is useful in conjunction with b).

d) Make sure master is ancestor of experimental or else format-patch
produces noisy set of experimental patches.

e) Fetch origin before doing anything so that any refspecs pointing
at origin are up-to-date.

f) Add ‘-n’ flag which prevents stuff from being sent to remote
repository.  All local operations are performed as usual.