Substitute %+character sequences in ‘tab-arguments’ resource
Due to backwards-incompatible changes ‘new-tab-command’ is no longer
supported.  This is a bit abrupt but it’s easier than having to maintain
backwards-compatilibity wrappers.  The hope is that not many people
actually used ‘new-tab-command’ and those that do will quickly figure
out what has happened.

The resource has been replaced with ‘tab-arguments’ which cana be used
to pass *any* arguments to the slave tab and supports some interpolation
for two-character sequences starting with a percent sign.

Most notably ‘%e’ is replaced by ‘-e <command>’ option passed when
starting urxvt and ‘%n’ is replaced by total number of tabs started
so far (excluding the one being created).

‘command-runner.sample’ scripts demonstrates how this can be used to
configure what commands are run in which tabs.