Revert "Support multiple values in ‘new-tab-command’ resource"
Instead of this I’m going to change the ‘new-tab-command’ resource
such that a sequence number can be passed to it.  With that, having
different commands depending which tab is created can be done by
users in a simple shell script.  This will simplify the tabbedex
plugin without removing any features.

This reverts commit 6e9d4e88fae74e873c36a01dae0aa6714284afe8.