Add support of changing directory to one in current tab
Introduce a ‘%p’ sequence to ‘tab-arguments’ resource which is replaced
by id of process group in foreground of the terminal in current tab.

With that change, command-runner.sample script can now be configured to
keep current working directory of the old tab when new tabs are created.
See command-runner.sample file for more information.