Add ‘%E’ sequence support to ‘tab-arguments’ resource
Introduce support for a ‘%E’ sequence in the ‘tab-arguments’
resource which is replaced by a command to execute.  It does not
include ‘-e’ prefix but at the same time is never empty defaulting
to shell if user did not specify ‘-e’ urxvt argument.

The sequence is perfect for executables which require a command to
run as last arguments, e.g.:

    URxvt.tabbedex.tab-arguments: -e nice %E