Rewrite command-runner.sample in shell
It’s likely that more people know shell than perl so having the sample
script written in the former should make it easier for people to set it
up to match their needs.

Furthermore, by using %E and getting rid of the code shifting identifier
when ‘-e’ is given, the script becomes much simpler which is another
added bonus.