Better error blocking in command-runner.sample script
Use trap to invoke the ‘Press RETURN to continue’ feature such that
the message is printed and script waits for user confirmation if the
exec call falls.  This is important since as soon as this script
terminates, tab in which it was started closes leaving user no chance
to read any error messages.