Support non-integer timeouts and make time-based rendering more accurate
The extension used to refresh the tabbar every second¹ to handle bell
and activity marks timeouts.  As a consequence, all of them were
accurate to within a second.  For example, if bell-timeout was one,
the bell could have been active for anything from 1.0 to 1.999…
seconds.

Change the code such that on each refresh, the next one is scheduled
at the moment when it is needed.

This has three consequences:
1. delays in bell-timeout and tabbar-timeouts no longer need to be
   whole numbers,
2. the timeouts are now much more precise and
3. refresh does not happen if there is nothing to update².

¹ Plus each time an event such as bell or new tab contents occurred.
² In other words, superfluous wake ups are removed which should help
  keep CPU in idle state.