Fix second terminal window hanging when urxvtd is used
tabbedex doesn’t relay on `on_foo’ subroutines being automatically
configured as hooks.  Instead, it explicitly enables during extension
initialisation.  To make this happen, when the perl module is loaded,
the code collects all hooks into a `%hooks` map which is then used by
an `enable_hooks` subroutine.  Since the dictionary is never needed
again, it is cleared at the end.

However, when urxvt is used, each perl extension is loaded only once
and all clients share the same global state.  Consequently, when the
‘%hooks’ hash is cleared, subsequent invocation of ‘enable_hooks’
subroutine are essentially noöps.

This results in a bug where second urxvt client hangs.

Fix this by not clearing the ‘%hooks’ mapping and keeping it in memory
for the next urxvtc to use.

Fixes: https://github.com/mina86/urxvt-tabbedex/issues/25