Add index support and ‘is_being_renamed’ check to make_current
Allow make_current function to be called with a numeric index in
addition to the tab.  This simplifies a common pattern of getting
a tab by index employed by many of the call sites.

Similarly, add check of whether the current tab is being renamed.
This again simplifies call sites which had to check for that and
makes the code less error-prone since now there’s no chance for
the check to be missed.

All this yields negative delta so yet another reason in favour of
this change.