Add ‘ignore_line_activity’ OSC which ignores updates to specified line
Normally tabbedex displays activity marks when new text is added to
a background tab.  The ‘tabbedex;ignore_line_activity’ OSC sequence
causes updates to specified line to be ignored.  This is useful when
application such as screen(1) has a clock in status line which updates
periodically without indicating real activity.

Fixes: https://github.com/mina86/urxvt-tabbedex/pull/35