Fix dead keys when using ibus and pointer is outside the window
When the mouse pointer is outside of the terminal window, dead keys
stop working.  When ibus is used, the prompt pops up but once next
key is pressed, no character is inserted to the terminal.  This fixes
that issue.

Unfortunately, the fix only works when ibus is used.  If urxvt is
configured not to utilise it, dead keys remain broken when the mouse
pointer is outside of the terminal window and result.

See: https://github.com/mina86/urxvt-tabbedex/issues/26
Fixes: https://github.com/mina86/urxvt-tabbedex/pull/39

[mina86@mina86.com: moved check to key_press handler and reworded
 comment and commit message]