Introduce `:nowrap` variants of `prev_tab` and `next_tab` actions
Add support for optional `:nowrap` argument to the `prev_tab` and
`next_tab` actions.  When used, switching tabs left or right will not
wrap around and instead stop at the first or last tab.  To use the
behaviour by default, the Shift-Left/Right bindings can be configured
to bind to the new actions:

    URxvt.keysym.Shift-Left:    tabbedex:prev_tab:nowrap
    URxvt.keysym.Shift-Right:   tabbedex:next_tab:nowrap

Fixes: https://github.com/mina86/urxvt-tabbedex/pull/42