Fix incorrect geometry when border or scroll bar is enabled
When border or scroll bar is enabled the -geometry flag isn’t quite
respected.  For example, starting the terminal via ‘urxvt -geometry
80x24 -pe tabbedex’ might result in internal tab having only 77
columns and 23 lines.  This is caused by root window being sized with
the assumption that there is no border or scroll bar even if they are
present in the tabs.

Fixes: https://github.com/mina86/urxvt-tabbedex/issues/44